Add zipkin tracing
Currently there's no way to trace the requests through the application.
This will add the tracing headers to any request that comes through this
application onto another service, or preserve existing headers allowing
the user to follow requests through the application as a whole. Zipkin
isds will also be added to logs to give better insight when debugging.